facturacion/pyproject.toml
[tool.poetry]
name = "facturacion"
version = "0.1.0"
description = ""
authors = ["fedexin40 <federico.cruz.lozano@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
psycopg2 = "^2.9.9"
satcfdi = "^4.6.0"
django-rest-framework = "^0.1.0"
gql = "^3.5.0"
aiohttp = "^3.10.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
